+++ src/ui/models/setmodel.h
@@ -7,7 +7,7 @@ class SetModel : public QObject
 {
     Q_OBJECT
     Q_PROPERTY(int repetitions READ repetitions WRITE setRepetitions NOTIFY repetitionsChanged)
-    Q_PROPERTY(int weight READ weight WRITE setWeight NOTIFY weightChanged)
+    Q_PROPERTY(double weight READ weight WRITE setWeight NOTIFY weightChanged)
     Q_PROPERTY(bool completed READ completed WRITE setCompleted NOTIFY completedChanged)
 
 public:
@@ -15,11 +15,11 @@ public:
     explicit SetModel(const Set &set, QObject *parent = nullptr);
 
     int repetitions() const;
-    int weight() const;
+    double weight() const;
     bool completed() const;
 
     void setRepetitions(int value);
-    void setWeight(int value);
+    void setWeight(double value);
     void setCompleted(bool value);
 
     const Set &entity() const;
